Catch local connection config
If database backend is Postgres and the connection is over domain
socket, the only values in <database> node are:
    * db_name
    * db_server
    * db_type

Now, for some reason the config reader unconditionally creates "tidy"
string representations for all possible keys. This means that
host, user and password are all empty strings (''), and not even
NoneType entities.

To catch the case for postgres, simply treat empty host the same as
undefined host.