PokerStarsTo Fpdb: Fix subtle bug in re_Action
The bet action could become multiline greedy if there was a playername starting with 'card' in the next line.

"""
danny purse: bets $0.50
cardjunkie25: calls $0.50
"""

Would match "danny purse: bets $0.50\ncard", and the rest of that line would be ignored.

Added '$' to the end of the regex to make sure that it is limited to 1 line, and wrap 'cards' as a group.