New: Tournament Results Viewer
Decided to see how hard it would be to get the tournament results using the existing schema.

Turns out the cash graph query is pretty similar and not much thought was required.

Created a new tab based on the existing cash game viewer for profit.

This is probably the idea page for multiple axis to be added.
- Distributions for position
- distributions by buy-in level
- ROI per buy-in level

Lots of stuff like that.