Add support for specifying DB port
In case someone needs to connect to a DB engine in non-standard port,
defining the TCP port is required. This is a rare corner case which
should only arise when a given DB host runs multiple versions of the
same DB engine.

However, the support is relatively straightforward to implement and it
shouldn't conflict with any other options.