Fix SVG deck path for installed software
When installing via setup.py we now include the SVG decks. As the path
is completely different, we also rewrite card_path.py to return a
build-time set path to the directory where all the decks are found.

Bump version number in setup.py while at it. The next release is 0.27
and these changes won't be in before that.