Stars: Remove ability to parse tourney results from hh file.
A recent patch allowing position to be added from Hand.select()
caused a crasher in tournament hands where the Stars parser attempted to
calculate tourney results from hh's.

I wasn't really happy with how the results parsing was being handled
originally, so i'm shifting the remaining code into an unused function
until the Importer is rewritten to use IdentifySite and gains the
ability to despatch text to hh parser, summary parser, or both.

I've shifted the old code into a new class method
readTourneyResults(self, hand) and made this function part of the HHC
API in anticipation of this happening.