Winamax: In cash games only hand.addPlayer() for players that acted.
Winamax cash hh's don't have a 'sitting out' section.

To fix the player list we split the hh into a header and summary, and
build a list of both sections.

If the lists are the same size, or it is a tourney we add the list of
players from the header.

If not we only add them from the summary section.